refactor: 플레이어 공통 로직 BasePlayer로 분리
- BasePlayer 클래스 생성: 렌더링, 물리엔진, 애니메이션 등 공통 로직 통합
- Player및 RemotePlayer가 BasePlayer를 상속받도록 구조 변경
