+++ frontend/src/game/players/RemotePlayer.ts
@@ -1,16 +1,7 @@
 import * as Phaser from "phaser";
-import { formatPlayTime } from "@/utils/timeFormat";
+import BasePlayer from "./BasePlayer";
 
-export default class RemotePlayer {
-  private scene: Phaser.Scene;
-  private bodySprite: Phaser.GameObjects.Sprite;
-  private container: Phaser.GameObjects.Container;
-  private body: Phaser.Physics.Arcade.Body;
-  private maskShape: Phaser.GameObjects.Graphics;
-  private faceSprite: Phaser.GameObjects.Image;
-  private timerText: Phaser.GameObjects.Text;
-  public id: string;
-
+export default class RemotePlayer extends BasePlayer {
   constructor(
     scene: Phaser.Scene,
     x: number,
@@ -19,88 +10,7 @@ export default class RemotePlayer {
     id: string,
     texture: string, // 유저네임(텍스처 키) 받기
   ) {
-    this.scene = scene;
-    this.id = id;
-
-    // 1. 컨테이너 생성
-    this.container = scene.add.container(x, y);
-
-    // 2. 몸통 스프라이트 생성
-    this.bodySprite = scene.add.sprite(0, 5, "body");
-    this.bodySprite.setFrame(0);
-
-    // 3. 얼굴 & 마스크
-    const FACE_RADIUS = 17;
-    const FACE_Y_OFFSET = 0;
-
-    this.maskShape = scene.make.graphics({});
-    this.maskShape.fillStyle(0xffffff);
-    this.maskShape.fillCircle(0, 0, FACE_RADIUS);
-    // 마스크 초기 위치를 컨테이너(플레이어) 위치로 설정
-    this.maskShape.x = x;
-    this.maskShape.y = y;
-    const mask = this.maskShape.createGeometryMask();
-
-    // 텍스처가 로드되어 있는지 확인하고, 없으면 기본값 'face'
-    const faceTexture = scene.textures.exists(texture) ? texture : "face";
-    this.faceSprite = scene.add.image(0, FACE_Y_OFFSET, faceTexture);
-
-    this.faceSprite.setDisplaySize(FACE_RADIUS * 2, FACE_RADIUS * 2);
-    this.faceSprite.setMask(mask);
-
-    // 4. 테두리
-    const borderGraphics = scene.add.graphics();
-    borderGraphics.lineStyle(2, 0xffffff, 1);
-    borderGraphics.strokeCircle(0, FACE_Y_OFFSET, FACE_RADIUS);
-
-    // 5. 닉네임 표시
-    const nameTag = scene.add
-      .text(0, 50, username, {
-        fontSize: "12px",
-        color: "#ffffff",
-        backgroundColor: "#00000088",
-        padding: { x: 4, y: 2 },
-      })
-      .setOrigin(0.5);
-
-    // 6. 접속 시간 표시
-    this.timerText = scene.add
-      .text(0, -35, formatPlayTime(0), {
-        fontSize: "12px",
-        color: "#ffffff",
-        fontFamily: "Arial, sans-serif",
-      })
-      .setOrigin(0.5);
-
-    this.timerText.setStroke("#000000", 4);
-    this.timerText.setShadow(1, 1, "#000000", 2, false, true);
-
-    // 7. 컨테이너 추가
-    this.container.add([
-      this.bodySprite,
-      this.faceSprite,
-      borderGraphics,
-      this.timerText,
-      nameTag,
-    ]);
-    this.container.setSize(FACE_RADIUS * 2, FACE_RADIUS * 3);
-
-    // 7. 물리 엔진 적용
-    scene.physics.world.enable(this.container);
-    this.body = this.container.body as Phaser.Physics.Arcade.Body;
-    this.body.setCollideWorldBounds(true);
-
-    this.body.setSize(FACE_RADIUS * 2, FACE_RADIUS * 3);
-    this.body.setOffset(0, 10);
-  }
-
-  // 얼굴 텍스처 업데이트
-  updateFaceTexture(texture: string) {
-    if (this.faceSprite && this.scene.textures.exists(texture)) {
-      this.faceSprite.setTexture(texture);
-      const FACE_RADIUS = 20;
-      this.faceSprite.setDisplaySize(FACE_RADIUS * 2, FACE_RADIUS * 2);
-    }
+    super(scene, x, y, username, id, texture);
   }
 
   // 서버에서 받은 상태로 업데이트
@@ -114,6 +24,9 @@ export default class RemotePlayer {
       console.error("RemotePlayer body not found");
       return;
     }
+
+    // 공통 update 호출 (마스크 동기화)
+    super.update();
 
     // 1. 위치 보정 (너무 멀어지면 강제 동기화)
     const dist = Phaser.Math.Distance.Between(
@@ -131,58 +44,31 @@ export default class RemotePlayer {
     this.container.y = Phaser.Math.Linear(this.container.y, state.y, 0.1);
 
     // 2. 속도 동기화 (움직임 반영)
-    const SPEED = 300;
     this.body.setVelocity(0); // 일단 정지
 
     if (state.isMoving) {
       // 방향 문자열 파싱해서 속도 적용
       if (state.direction.includes("left")) {
-        this.body.setVelocityX(-SPEED);
-        this.bodySprite.play("walk-left", true);
+        this.body.setVelocityX(-this.speed);
+        this.playAnimation("walk-left");
       }
       if (state.direction.includes("right")) {
-        this.body.setVelocityX(SPEED);
-        this.bodySprite.play("walk-right", true);
+        this.body.setVelocityX(this.speed);
+        this.playAnimation("walk-right");
       }
       if (state.direction.includes("up")) {
-        this.body.setVelocityY(-SPEED);
-        this.bodySprite.play("walk-up", true);
+        this.body.setVelocityY(-this.speed);
+        this.playAnimation("walk-up");
       }
       if (state.direction.includes("down")) {
-        this.body.setVelocityY(SPEED);
-        this.bodySprite.play("walk-down", true);
+        this.body.setVelocityY(this.speed);
+        this.playAnimation("walk-down");
       }
     } else {
       // 멈췄을 때
-      this.bodySprite.stop();
+      this.stopAnimation();
       // 멈췄을 때는 좌표 강제 동기화 (정확한 위치 안착)
       this.container.setPosition(state.x, state.y);
     }
-
-    // 마스크 동기화 (update()에서 수행하므로 여기서는 제거해도 되지만, 안전을 위해 둠)
-    if (this.maskShape) {
-      this.maskShape.x = this.container.x;
-      this.maskShape.y = this.container.y;
-    }
-  }
-
-  // 매 프레임 호출되어 마스크 위치를 컨테이너와 동기화
-  update() {
-    if (this.container && this.maskShape) {
-      this.maskShape.x = this.container.x;
-      this.maskShape.y = this.container.y;
-    }
-  }
-
-  // 접속 시간 업데이트
-  updateTimer(minutes: number) {
-    if (this.timerText) {
-      this.timerText.setText(formatPlayTime(minutes));
-    }
-  }
-
-  destroy() {
-    this.container.destroy();
-    this.maskShape.destroy();
   }
 }
